method/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "operator-method"
version = "0.1.0"
description = "Zero-loss context compaction for AI agents using parallel fast-inference memory rescue"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "the operator", email = "josh@operator.com"}]
requires-python = ">=3.10"
keywords = ["ai", "agents", "memory", "context", "compaction", "llm", "qdrant", "vector-database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "httpx>=0.25.0",
]

[project.optional-dependencies]
qdrant = ["qdrant-client>=1.7.0"]
chroma = ["chromadb>=0.4.0"]
pinecone = ["pinecone-client>=3.0.0"]
all = ["qdrant-client>=1.7.0", "chromadb>=0.4.0", "pinecone-client>=3.0.0"]

[project.urls]
Homepage = "https://github.com/joperator/operator-method"
Repository = "https://github.com/joperator/operator-method"
Issues = "https://github.com/joperator/operator-method/issues"

[tool.setuptools.packages.find]
where = ["src"]
